Make Wrath Server header size be u32
Wrath introduces a special case where 3 bytes can be used for size for
server messages instead of just 2.

https://github.com/mangostwo/server/blob/1a4aa1fccfea239ed1dcba81bf6fe10543b8a124/src/game/Server/WorldSocket.cpp#L61

+++ src/wrath_header/mod.rs
@@ -34,7 +34,7 @@ const R: [u8; 16] = [
 
 #[derive(Debug, Clone, Copy)]
 pub struct ServerHeader {
-    pub size: u16,
+    pub size: u32,
     pub opcode: u16,
 }
